add audio user evaluation with API and fix audio recording

+++ app/src/main/java/com/bangkit/hijalearn/data/remote/response/PredictionResponse.kt
@@ -8,7 +8,7 @@ data class PredictionResponse(
 	val correct: Boolean,
 
 	@field:SerializedName("probability")
-	val probability: Int,
+	val probability: Double,
 
 	@field:SerializedName("message")
 	val message: String
